Move partial-colorScale tests into colorScaleMissingData.test.ts
Consolidate into the existing cross-series colorScale edge-case suite
rather than adding a standalone file.
