+++ packages/ag-charts-enterprise/src/test/colorScaleMissingData.test.ts
@@ -1,4 +1,4 @@
-import { afterEach, describe, it } from '@jest/globals';
+import { afterEach, describe, expect, it } from '@jest/globals';
 
 import {
     type AgBubbleSeriesOptions,
@@ -15,6 +15,8 @@ import {
     waitForChartStability,
 } from 'ag-charts-community-test';
 
+import { ukData } from '../series/map-test/ukData';
+import ukTopology from '../series/map-test/ukTopology.json';
 import { prepareEnterpriseTestOptions } from './utils';
 
 interface ColorScaleSeriesLike {
@@ -146,4 +148,134 @@ describe('colorScale.missingDataFill - bubble/scatter', () => {
             expect(capturedFill).toBe(missingDataFill);
         }
     );
+});
+
+// AG-17296: Verify that a user-supplied partial `colorScale` (e.g. `{}` or `{ mode: 'discrete' }`)
+// does not wipe the theme-supplied fills for series that derive their palette via `$map` theme
+// expressions.
+describe('colorScale partial options — theme fills survive user partials', () => {
+    setupMockConsole();
+    setupMockCanvas();
+
+    let chart: any;
+
+    afterEach(() => {
+        if (chart) {
+            chart.destroy();
+            chart = undefined;
+        }
+    });
+
+    function assertColorScalePopulated() {
+        const series: any = deproxy(chart).series.find((s: any) => s.colorScale != null);
+        expect(series).toBeDefined();
+        expect(series.colorScale.domain.length).toBeGreaterThanOrEqual(2);
+        expect(series.colorScale.range.length).toBeGreaterThanOrEqual(2);
+        expect(series.colorScale.range).not.toEqual(['red', 'blue']);
+    }
+
+    it('heatmap with colorScale: {}', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: [
+                { x: 'A', y: '1', v: 10 },
+                { x: 'B', y: '1', v: 20 },
+                { x: 'A', y: '2', v: 30 },
+                { x: 'B', y: '2', v: 40 },
+            ],
+            series: [{ type: 'heatmap', xKey: 'x', yKey: 'y', colorKey: 'v', colorScale: {} }],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
+
+    it('treemap with colorScale: {}', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: [
+                { title: 'A', size: 10, color: 1, children: [{ title: 'A1', size: 5, color: 2 }] },
+                { title: 'B', size: 20, color: 3, children: [{ title: 'B1', size: 8, color: 4 }] },
+            ],
+            series: [
+                {
+                    type: 'treemap',
+                    labelKey: 'title',
+                    sizeKey: 'size',
+                    colorKey: 'color',
+                    childrenKey: 'children',
+                    colorScale: {},
+                },
+            ],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
+
+    it('sunburst with colorScale: {}', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: [
+                { name: 'A', size: 10, color: 1, children: [{ name: 'A1', size: 5, color: 2 }] },
+                { name: 'B', size: 20, color: 3, children: [{ name: 'B1', size: 8, color: 4 }] },
+            ],
+            series: [
+                {
+                    type: 'sunburst',
+                    labelKey: 'name',
+                    sizeKey: 'size',
+                    colorKey: 'color',
+                    childrenKey: 'children',
+                    colorScale: {},
+                },
+            ],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
+
+    it('map-marker with colorScale: {}', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: ukData,
+            topology: ukTopology,
+            series: [
+                { type: 'map-shape-background' },
+                { type: 'map-marker', idKey: 'name', colorKey: 'population', colorScale: {} },
+            ],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
+
+    it('map-shape with colorScale: {}', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: ukData,
+            topology: ukTopology,
+            series: [{ type: 'map-shape', idKey: 'name', colorKey: 'population', colorScale: {} }],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
+
+    it('heatmap with colorScale: { mode: "discrete" }', async () => {
+        const options: AgChartOptions = prepareEnterpriseTestOptions({
+            data: [
+                { x: 'A', y: '1', v: 10 },
+                { x: 'B', y: '1', v: 20 },
+                { x: 'A', y: '2', v: 30 },
+                { x: 'B', y: '2', v: 40 },
+            ],
+            series: [{ type: 'heatmap', xKey: 'x', yKey: 'y', colorKey: 'v', colorScale: { mode: 'discrete' } }],
+        });
+
+        chart = AgCharts.create(options);
+        await waitForChartStability(chart);
+        assertColorScalePopulated();
+    });
 });
